+++ src/grafv4/generic_counter.rs
