+++ src/grafv4/mod.rs
@@ -1,6 +1,6 @@
 pub(crate) mod count_mode;
 pub(crate) mod description_printer;
-mod graf_zahl;
+mod generic_counter;
 mod io_reader;
 mod countable;
 mod count_mode_lib;
